+++ dpdocter/src/main/java/com/dpdocter/beans/InventoryItem.java
@@ -117,11 +117,11 @@ public class InventoryItem extends GenericCollection {
 		this.reOrderLevel = reOrderLevel;
 	}
 
-	public Long getRetailPrice() {
+	public Double getRetailPrice() {
 		return retailPrice;
 	}
 
-	public void setRetailPrice(Long retailPrice) {
+	public void setRetailPrice(Double retailPrice) {
 		this.retailPrice = retailPrice;
 	}
